Added an ELSE action
Currently, actions are executed only when the condition is evaluated to be true. However, from time to time, the rule engine needs to do some other actions when the condition is false. In MVEL, or Easy Rule, this can only be achieved by creating another rule that has the reverse logic, so the other actions can be executed. The ELSE action will allow the rule authors to add other actions (ELSE actions) within the same rule.